Database initialization - check if database exists before resetting

+++ src/db/init.sh
@@ -1,7 +1,15 @@
-#run with 'bash filename'
+#!/bin/bash
 
-dropdb sundial
-createdb sundial
-psql -d sundial < ./ddl/monitor.sql
-psql -d sundial < ./ddl/run.sql
-psql -d sundial < ./ddl/ping.sql
+# Unless the sundial database exists and contains both monitor, run and ping tables, reset database
+
+if ! $(psql -lqt | cut -d \| -f 1 | grep -qw "sundial" && \
+       psql -d "sundial" -c "\dt" | grep -q "monitor" && \
+       psql -d "sundial" -c "\dt" | grep -q "ping" && \
+       psql -d "sundial" -c "\dt" | grep -q "run" ); then
+
+  dropdb sundial
+  createdb sundial
+  psql -d sundial < ./ddl/monitor.sql
+  psql -d sundial < ./ddl/run.sql
+  psql -d sundial < ./ddl/ping.sql
+fi
